Replace hack from #2354 with flexbox solution
The stage + target pane wrapper's width must be constrained to that of
the stage, even when there are lots of sprite tiles. Otherwise, once you
add enough sprites, it'll grow horizontally past the stage size. The old
code set its maximum width explicitly to stop that from happening, but
it can also be done more elegantly with the flex-basis property.
This also has the benefit of taking the stage border width into account.