allow 'About' button to open a menu
If `onClickAbout` is a callback, or is not provided, then behavior is
the same as it was before this change. If the value is an array of
objects with `title` and `onClick` properties, the "About" button will
now open a menu with those items.

TODO: support `FormattedMessage` for the menu item titles.